feat: start moving AuthContext hooks and utils to wallet package
