--- src/context/AuthContext/useSigner/__mocks__/index.ts
@@ -1,4 +1,0 @@
-import fakeSigner from '__mocks__/models/signer';
-
-const useSigner = () => fakeSigner;
-export default useSigner;
