+++ src/packages/contracts/scripts/generateContractRecords/generateContracts/generateGetters/templates/uniqueContractGettersTemplate.hbs
@@ -7,9 +7,8 @@ import { useMemo } from 'react';
 import { ChainId } from 'types';
 
 import { useAuth } from 'context/AuthContext';
-import useProvider from 'context/AuthContext/useProvider';
-import useSigner from 'context/AuthContext/useSigner';
 import { getUniqueContractAddress } from 'packages/contracts/utilities/getUniqueContractAddress';
+import { useProvider, useSigner } from 'packages/wallet';
 
 interface Get{{contractName}}ContractAddressInput {
   chainId: ChainId;
