+++ src/packages/wallet/Web3Wrapper/__mocks__/index.tsx
@@ -2,6 +2,4 @@ import React from 'react';
 
 import { Web3WrapperProps } from '..';
 
-const Web3Wrapper: React.FC<Web3WrapperProps> = ({ children }) => <>{children}</>;
-
-export default Web3Wrapper;
+export const Web3Wrapper: React.FC<Web3WrapperProps> = ({ children }) => <>{children}</>;
