+++ src/packages/wallet/hooks/useProvider/__mocks__/index.ts
@@ -1,5 +1,4 @@
 import { getDefaultProvider } from 'ethers';
 
 const defaultProvider = getDefaultProvider();
-const useProvider = () => defaultProvider;
-export default useProvider;
+export const useProvider = () => defaultProvider;
