+++ src/packages/wallet/hooks/useProvider/index.ts
@@ -2,15 +2,13 @@ import { useMemo } from 'react';
 import { ChainId } from 'types';
 import { usePublicClient } from 'wagmi';
 
-import getProvider from './getProvider';
+import { getProvider } from './getProvider';
 
 export interface UseProviderInput {
   chainId?: ChainId;
 }
 
-const useProvider = (input?: UseProviderInput) => {
+export const useProvider = (input?: UseProviderInput) => {
   const publicClient = usePublicClient({ chainId: input?.chainId });
   return useMemo(() => getProvider({ publicClient }), [publicClient]);
 };
-
-export default useProvider;
