+++ src/packages/wallet/hooks/useSigner/__mocks__/index.ts
@@ -1,0 +1,3 @@
+import fakeSigner from '__mocks__/models/signer';
+
+export const useSigner = () => fakeSigner;
