+++ src/packages/wallet/hooks/useSigner/index.ts
@@ -2,15 +2,13 @@ import { useMemo } from 'react';
 import { ChainId } from 'types';
 import { useWalletClient } from 'wagmi';
 
-import getSigner from './getSigner';
+import { getSigner } from './getSigner';
 
 export interface UseSignerInput {
   chainId?: ChainId;
 }
 
-const useSigner = (input?: UseSignerInput) => {
+export const useSigner = (input?: UseSignerInput) => {
   const { data: walletClient } = useWalletClient({ chainId: input?.chainId });
   return useMemo(() => getSigner({ walletClient: walletClient || undefined }), [walletClient]);
 };
-
-export default useSigner;
