+++ src/packages/wallet/index.ts
@@ -1,7 +1,8 @@
 export * from './connectors';
 export * from './types';
-export { default as Web3Wrapper } from './Web3Wrapper';
 export * from './Web3Wrapper';
 export * from './chains';
-export * from './canAddTokenToWallet';
-export * from './addTokenToWallet';
+export * from './utilities/canAddTokenToWallet';
+export * from './utilities/addTokenToWallet';
+export * from './hooks/useProvider';
+export * from './hooks/useSigner';
