+++ src/packages/wallet/utilities/addTokenToWallet/index.ts
