+++ src/packages/wallet/utilities/canAddTokenToWallet/index.ts
